Concatenate operators

+++ app/src/main/java/com/globant/counter/android/operators/Division.java
@@ -2,7 +2,7 @@ package com.globant.counter.android.operators;
 
 public class Division extends Operator {
 
-    public float operate (int x, int y) {
+    public float operate (float x, float y) {
         if (y == 0) {
             return 666;
         }
