+++ app/src/main/java/com/globant/counter/android/operators/Multiplication.java
@@ -3,7 +3,7 @@ package com.globant.counter.android.operators;
 
 public class Multiplication extends Operator {
 
-    public float operate (int x, int y) {
+    public float operate (float x, float y) {
         return x * y;
     }
