+++ app/src/main/java/com/globant/counter/android/operators/Operator.java
@@ -2,6 +2,6 @@ package com.globant.counter.android.operators;
 
 public abstract class Operator {
 
-    public abstract float operate (int x, int y);
+    public abstract float operate (float x, float y);
 
 }
